Keyphrase extraction
git-svn-id: https://141.76.40.86/svn-students/iircommon/toolkit/trunk@1266 5654db93-19e5-4567-beb8-64e8c25c7c45

+++ src/main/java/tud/iir/classification/controlledtagging/CandidateComparator.java
@@ -7,9 +7,7 @@ public class CandidateComparator implements Comparator<Candidate> {
 
     @Override
     public int compare(Candidate cand0, Candidate cand1) {
-        return new Float(cand1.getRegressionValue()).compareTo(cand0.getRegressionValue());
-        //return new Float(cand1.getTermFrequencyInverseDocumentFrequency()).compareTo(cand0.getTermFrequencyInverseDocumentFrequency());
-
+        return new Double(cand1.getRegressionValue()).compareTo(cand0.getRegressionValue());
     }
 
 }
